Moved from service package
