--- src/main/java/br/com/pipocaagil/apipipocaagil/payments/representations/AmountRepresentation.java
@@ -1,3 +1,0 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
-
-public record AmountRepresentation(Integer value, String currency) {}
