--- src/main/java/br/com/pipocaagil/apipipocaagil/payments/representations/HolderRepresentation.java
@@ -1,3 +1,0 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
-
-public record HolderRepresentation(String name) {}
