--- src/main/java/br/com/pipocaagil/apipipocaagil/payments/representations/ShippingRepresentation.java
@@ -1,4 +1,0 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
-
-public record ShippingRepresentation(AddressRepresentation address) {
-}
