+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/exception/HttpClientErrorException.java
@@ -1,4 +1,4 @@
-package br.com.pipocaagil.apipipocaagil.payments.exception;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.exception;
 
 public class HttpClientErrorException extends RuntimeException{
     public HttpClientErrorException(String message) {
