+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/exception/JsonProcessingException.java
@@ -1,4 +1,4 @@
-package br.com.pipocaagil.apipipocaagil.payments.exception;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.exception;
 
 public class JsonProcessingException extends RuntimeException{
     public JsonProcessingException(String message) {
