+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/AddressRepresentation.java
@@ -1,4 +1,4 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
 
 public record AddressRepresentation(String street, String number, String complement, String locality, String city, String region_code,
                                     String country, String postal_code) {}
