+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/ChargeRepresentation.java
@@ -1,3 +1,3 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
 
 public record ChargeRepresentation(AmountRepresentation amount, PaymentMethodRepresentation payment_method, String reference_id, String description) {}
