+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/HolderRepresentation.java
@@ -1,0 +1,3 @@
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
+
+public record HolderRepresentation(String name) {}
