+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/ItemsRepresentation.java
@@ -1,3 +1,3 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
 
 public record ItemsRepresentation(String name, Integer quantity, Integer unit_amount) {}
