+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/PaymentMethodRepresentation.java
@@ -1,3 +1,3 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
 
 public record PaymentMethodRepresentation(CardRepresentation card, String type, Integer installments, Boolean capture, String soft_descriptor) {}
