+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/PhonesRepresentation.java
@@ -1,3 +1,3 @@
-package br.com.pipocaagil.apipipocaagil.payments.representations;
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
 
 public record PhonesRepresentation(Integer country, Integer area, Integer number, String type) {}
