+++ src/main/java/br/com/pipocaagil/apipipocaagil/services/paymentservice/representations/ShippingRepresentation.java
@@ -1,0 +1,4 @@
+package br.com.pipocaagil.apipipocaagil.services.paymentservice.representations;
+
+public record ShippingRepresentation(AddressRepresentation address) {
+}
